+++ fintech/src/main/java/org/weather/client/RestTemplateConfig.java
@@ -1,4 +1,4 @@
-package org.weather.config;
+package org.weather.client;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
