+++ fintech/src/main/resources/application.properties
@@ -21,4 +21,4 @@ spring.jpa.hibernate.ddl-auto=validate
 
 spring.h2.console.enabled=true
 
-hibernate.enabled=false
+hibernate.enabled=true
