+++ fintech/src/main/resources/db/changelog/changeset/add-city-date-composite-index-to-weather.yaml
